refactor(mcp): extract skardi-mcp-core — shared projection + encode_component
The MCP HTTP transport spec (docs/superpowers/specs/
2026-08-28-mcp-http-transport-design.md) names two consumers of the
projection layer, so the drift-prone knowledge moves to a shared crate:
projection.rs wholesale (schema generation, sanitize/assign_tool_names,
built-in constants, version-skew fallbacks, all unit tests) and
encode_component (part of the tool->REST translation contract). The CLI
re-exports encode_component from crate::client so call sites are
unchanged; the dispatch matches deliberately stay per-binding.

Also updates the CLI's rmcp comment that feature unification falsifies
once the server grows rmcp's HTTP stack.
