+++ crates/cli/src/client.rs
@@ -4,7 +4,6 @@
 //! responses (and transport failures) into a uniform `ApiError`.
 
 use crate::config::ClientConfig;
-use percent_encoding::{AsciiSet, NON_ALPHANUMERIC, utf8_percent_encode};
 use reqwest::header::{AUTHORIZATION, HeaderName, HeaderValue};
 use reqwest::{Client, RequestBuilder, Response, StatusCode};
 use serde_json::Value;
@@ -12,16 +11,10 @@ use std::error::Error as StdError;
 use std::fmt;
 use std::net::{Ipv4Addr, Ipv6Addr};
 
-/// Characters percent-encoded by [`encode_component`]: everything except
-/// ASCII alphanumerics and the RFC 3986 "unreserved" marks (`-`, `.`, `_`,
-/// `~`). Deliberately conservative — over-encoding is always valid, while
-/// missing a reserved character (`/`, `?`, `#`, `%`, space, …) mis-routes
-/// the request.
-const URL_COMPONENT: &AsciiSet = &NON_ALPHANUMERIC
-    .remove(b'-')
-    .remove(b'.')
-    .remove(b'_')
-    .remove(b'~');
+// The pipeline-name → URL path encoding is part of the tool→REST translation
+// contract shared with the server's /mcp binding; the CLI re-exports it so
+// command modules keep importing from crate::client.
+pub use skardi_mcp_core::encode_component;
 
 /// [`WORKSPACE_HEADER`] lowercased, for `HeaderName::from_static` — which
 /// panics on any uppercase byte, and is the only constructor that cannot fail
@@ -56,13 +49,6 @@ fn retry_after_seconds(response: &Response) -> Option<u64> {
         .trim()
         .parse()
         .ok()
-}
-
-/// Percent-encode one URL path segment or query value (user-supplied
-/// pipeline/job names, run ids) so characters like `/`, `?`, `#`, `%`, and
-/// spaces cannot alter the request route.
-pub fn encode_component(raw: &str) -> String {
-    utf8_percent_encode(raw, URL_COMPONENT).to_string()
 }
 
 /// A thin HTTP client for a single skardi-server instance.
@@ -584,22 +570,6 @@ mod tests {
     fn base_url_strips_trailing_slash() {
         let client = ApiClient::new(&config("http://h:1/", None)).unwrap();
         assert_eq!(client.base_url, "http://h:1");
-    }
-
-    #[test]
-    fn encode_component_escapes_reserved_and_keeps_unreserved() {
-        use super::encode_component;
-
-        assert_eq!(encode_component("a/b"), "a%2Fb");
-        assert_eq!(encode_component("a b?c#d%e"), "a%20b%3Fc%23d%25e");
-        assert_eq!(encode_component("a&b=c"), "a%26b%3Dc");
-        // Unreserved characters pass through untouched.
-        assert_eq!(
-            encode_component("daily-report_v2.1~x"),
-            "daily-report_v2.1~x"
-        );
-        // Non-ASCII is UTF-8 percent-encoded.
-        assert_eq!(encode_component("café"), "caf%C3%A9");
     }
 
     #[test]
